pyproject.toml
[tool.poetry]
name = "hospital-system"
version = "0.1.0"
description = ""
authors = ["denys-stefanko <partner.denys.stefanko@philips.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.1.0"
flask-sqlalchemy = "^3.1.1"
flask-migrate = "^4.0.7"
psycopg2-binary = "^2.9.10"
pydantic = "^2.10.4"
google-generativeai = "^0.8.3"
python-dotenv = "^1.0.1"
dependency-injector = "^4.44.0"
yagmail = "^0.15.293"
flask-mail = "^0.10.0"
pip = "^24.3.1"
google-auth = "^2.37.0"
google-auth-oauthlib = "^1.2.1"
google-auth-httplib2 = "^0.2.0"
google-api-python-client = "^2.157.0"

[tool.poetry.dev-dependencies]
pytest = "^8.3.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
